+++ CHANGES.rst
@@ -1,12 +1,14 @@
 Changelog
 =========
 
-Version 0.10.1 (bugfixes, not yet released)
+Version 0.11 (not yet released)
 -------------------------------------------
 
 - Update dependencies (old GitPython was broken).
 - Fix up documentation.
 - JIRA issue fetching normalizes ticket IDs to be uppercase. See #84.
+- Update styler format to allow better grouping and ordering.
+- Add support for docformatter styler. See #106.
 
 Version 0.10.0 (Released Jul 2, 2017)
 -------------------------------------